# swagger.yaml
openapi: 3.0.0
info:
  title: AI Power Watch API
  description: An AI-powered E-commerce website using the MERN stack
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
paths:
...
  "/users":
    get:
      security:
        - bearerAuth: []
      summary: Get all users
      tags: [User]
      parameters:
        - $ref: '#/components/parameters/pageParams'
        - $ref: '#/components/parameters/limitParams'
        - $ref: '#/components/parameters/sortByParams'
        - $ref: '#/components/parameters/sortTypeParams'
        - $ref: '#/components/parameters/searchParams'
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDTO'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        example: "/users"
                      nextPage:
                        type: string
                        example: "/users?page=1&limit=10"
                      prevPage:
                        type: string
                        example: "/users?page=3&limit=10"
        '500':
          $ref: '#/components/responses/500'
        '400':
          $ref: '#/components/responses/400'
  "/users/{id}":
    get:
      security:
        - bearerAuth: []
      summary: Get a user
      tags: [User]
      parameters:
        - $ref: '#/components/parameters/idParams'
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Get Your user successfully
                  data:
                    $ref: '#/components/schemas/UserDTO'
                  
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        
    put:
      security:
        - bearerAuth: []
      summary: Update a user
      tags: [User]
      parameters:
        - $ref: '#/components/parameters/idParams'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Update A user successfully
                  data:
                    $ref: '#/components/schemas/UserDTO'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      security:
        - bearerAuth: []
      summary: Delete a user
      tags: [User]
      parameters:
        - $ref: '#/components/parameters/idParams'
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Delete A user successfully
                  data:
                    $ref: '#/components/schemas/UserDTO'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Responses
  responses:
...
    "500":
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the inconvenience. Please try again later.
    "400":
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              message:
                type: string
                example: Bad request
...
    "404":
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              message:
                type: string
                example: Resource Not Found
  # Schemas
  schemas:
...
    UserDTO:
      type: object
      properties:
        name:
          type: string
          example: John
        email:
          type: string
          example: 0l6g6@example.com
        role:
          type: string
          example: admin
        avatar:
          type: string
          example: https://www.gravatar.com/avatar/205e460b479e2e5b48aec07710c08d50
      required:
      - name
      - email

    # @ Product Item DTO
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        limit:
          type: integer
          format: int32
          example: 10
        next:
          type: integer
          format: int32
          example: 3
        prev:
          type: integer
          format: int32
          example: 1
        totalItems:
          type: integer
          format: int32
          example: 25
        totalPage:
          type: integer
          format: int32
          example: 3
      required: [page, limit, totalItems, totalPage]
    
  # Parameters
  parameters:
    pageParams:
      in: query
      name: page
      description: The page number
      schema:
        type: integer
        format: int32
        example: 1
    limitParams:
      in: query
      name: limit
      description: The number of items per page should be return
      schema:
        type: integer
        format: int32
        example: 10
    idParams:
      in: path
      name: id
      description: The id of the Item
      required: true
      schema:
        type: string
        example: 63d0e1b5b0e6b1b0e6b1b0e6
    sortByParams:
      in: query 
      name: sort_by
      description: The field to sort by (price, rating, stock, name, category)
      schema:
        type: string
        example: price
    sortTypeParams:
      in: query
      name: sort_type
      description: The sort type (asc, desc)
      schema:
        type: string
        enum: [asc, desc]
        example: asc
    searchParams:
      in: query
      name: search
      description: Search by product name
      schema:
        type: string
        example: ""